Previous Update issues:- After adding price(user input) to EditText in each itemview scrolling the view lead to either losing values or is seen in EdiText of other itemview.
Analysis:- Since each EditText (Price) had textchange listener assigned to it, whenever the RecyclerView is scrolled BindView in UpdateRecyclerAdapter is called. Even though it should work in logical sense, since the recyclerview recycles the view holder.getAdapterPosition() won't return desired value. Hence the values in EditText is lost or get jumbled.

Fix:- Instead of using TextChangedlistener in BindView or ViewHolder directly each EditText(Price) has focuschangedlistener, which has boolean value based on whether the EditText is focused or not. If focused that EditText is assigned a TextChangedlistener and we can do the operations there. If it lose the focus we remove TextChangedListener from that EditText hence scrolling won't affect these views. So at a particular instant either a single edittext has TextChangedlistener or none of the edittext do.

Minor Update:- Added a function in DataBaseHandler to add price to the particular item no.

                            In manifest added windowsoftinputmode for hiding softkeyboard when application is opened and adjust the view based on where the EditText is. Previously for the EditText present at lower screen the softkeyboard will cover the edittext and will lose focus. Hence not being able to edit the field.